+++ packages/vuikit/src/library/spinner/elements/index.js
@@ -1,1 +1,2 @@
 export { default as ElementSpinner } from './spinner'
+export { default as ElementSpinnerIcon } from './spinner--icon'